+++ bin-proto-derive/src/codegen/enums.rs
@@ -3,11 +3,8 @@ use crate::{
     codegen,
     plan::{self, EnumVariant},
 };
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 
-/// Generates code that reads one of a set of
-/// parcel variants and returns an expression
-/// of the same type as the enum.
 pub fn write_variant(
     plan: &plan::Enum,
     write_discriminant: &dyn Fn(&EnumVariant) -> TokenStream,
@@ -17,11 +14,8 @@ pub fn write_variant(
         .iter()
         .map(|variant| {
             let variant_name = &variant.ident;
-
             let write_discriminant = write_discriminant(variant);
-
             let fields_pattern = bind_fields_pattern(variant_name, &variant.fields);
-
             let writes = codegen::writes(&variant.fields, false);
 
             quote!(Self :: #fields_pattern => {
@@ -34,7 +28,6 @@ pub fn write_variant(
     quote!(
         match *self {
             #(#variant_match_branches,)*
-            _ => return Err(bin_proto::Error::UnknownEnumDiscriminant(String::new())),
         }
     )
 }
@@ -44,11 +37,11 @@ pub fn read_variant(
     read_discriminant: TokenStream,
     attribs: &Attrs,
 ) -> TokenStream {
-    let discriminant_ty = plan.discriminant_ty.clone();
+    let discriminant_ty = &plan.discriminant_ty;
 
     let discriminant_match_branches = plan.variants.iter().map(|variant| {
         let variant_name = &variant.ident;
-        let discriminant_literal = variant.discriminant_value.clone();
+        let discriminant_literal = &variant.discriminant_value;
         let (reader, initializer) = codegen::reads(&variant.fields, attribs);
 
         quote!(
@@ -66,8 +59,8 @@ pub fn read_variant(
             match discriminant {
                 #(#discriminant_match_branches,)*
                 unknown_discriminant => {
-                    return Err(bin_proto::Error::UnknownEnumDiscriminant(
-                        format!("{:?}", unknown_discriminant),
+                    return Err(::bin_proto::Error::UnknownEnumDiscriminant(
+                        ::std::format!("{:?}", unknown_discriminant),
                     ));
                 },
             }
@@ -75,9 +68,6 @@ pub fn read_variant(
     )
 }
 
-/// Generates code for a pattern that binds a set of fields by reference.
-///
-/// Returns a tuple of the pattern tokens and the field binding names.
 pub fn bind_fields_pattern(parent_name: &syn::Ident, fields: &syn::Fields) -> TokenStream {
     match *fields {
         syn::Fields::Named(ref fields_named) => {
@@ -92,7 +82,7 @@ pub fn bind_fields_pattern(parent_name: &syn::Ident, fields: &syn::Fields) -> To
         }
         syn::Fields::Unnamed(ref fields_unnamed) => {
             let binding_names: Vec<_> = (0..fields_unnamed.unnamed.len())
-                .map(|i| syn::Ident::new(&format!("field_{}", i), proc_macro2::Span::call_site()))
+                .map(|i| syn::Ident::new(&format!("field_{}", i), Span::call_site()))
                 .collect();
 
             let field_refs: Vec<_> = binding_names.iter().map(|i| quote!( ref #i )).collect();