+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # v0.4.0
 - Delete `EnumExt`
-- Bump rust version to 2021
+- Bump dependencies, and rust version to 2021
 - Make lifetime generics work
 - Handle context using generics instead of `Any`
+- Improve derive macro hygiene
+- Improve derive macro error reporting
 # v0.3.4
 - Do not trigger https://github.com/rust-lang/rust/issues/120363 with generated code
 # v0.3.3